consolidate checks
